refactor(chat): split ChatAnimation into smaller components and extract animation logic to custom hook
