+++ src/components/Community/ChatAnimation/constants/messages.data.ts
@@ -1,0 +1,28 @@
+export const messages = [
+  {
+    user: "Anna",
+    avatar: "👩‍💻",
+    message: "Hej! Ktoś może pomóc z Promise w JS?",
+    time: "12:01"
+  },
+  {
+    user: "Michał",
+    avatar: "👨‍💻",
+    message: "Jasne! Z czym dokładnie masz problem?",
+    time: "12:02"
+  },
+  {
+    user: "Anna",
+    avatar: "👩‍💻",
+    message: "Nie do końca rozumiem async/await...",
+    time: "12:03"
+  },
+  {
+    user: "Tomek",
+    avatar: "🧑‍💻",
+    message: "Mogę pokazać ci kilka przykładów!",
+    time: "12:04"
+  }
+];
+
+export const typingText = "Dziękuję za pomoc! 🙏";
