+++ src/components/Community/ChatAnimation/types/types.ts
@@ -1,0 +1,6 @@
+export type Message = {
+  user: string;
+  avatar: string;
+  message: string;
+  time: string;
+};
